Cyberiad (YogStation copy) (#22)
<!-- If this is your first PR, or not, take the time to read our
CONTRIBUTING.md file! You can see it here:
https://github.com/yogstation13/Yogstation/blob/master/.github/CONTRIBUTING.md
You can remove all headers (Document the changes, Spriting and Wiki
documentation) if there is no wiki documentation required but you must
still explain what the pr is and why it needs to be added to the game.
Directors+ Are immune from this rule in exceptional circumstances. -->

# Document the changes in your pull request
Скопировал YogStation для работы без конфликтов в будущем
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Prefix the PR title with [admin] if it involves something admin related.
Prefix the PR title with [s] if you are fixing an exploit, so that it is
not announced on the Yogstation Discord and the server.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without
saying.-->

# Changelog

<!-- Edit the changelog below to reflect the changes made by this PR,
even if the changes are minor - required for every PR that has
player-facing changes.
If you add a name after the ':cl:', that name will be used in the
changelog. Leave it empty to use your GitHub name. -->

:cl:  
add: Cyberiad map
/:cl: